Create mailsenderv2.py
The "credentials.json" file used in the example code above is a JSON file that stores the OAuth 2.0 authentication client credentials you created in the Google Cloud Console. You need to specify this file as the value of the "SERVICE_ACCOUNT_FILE" variable in the sample code above. You also need to change the "to", "subject" and "body" variables to reflect the recipient, subject and content of your message.